cloneFinder/requirements.txt
requests
pyyaml
neo4j